--- a/lib/compress/zstd_compress_internal.h
+++ b/lib/compress/zstd_compress_internal.h
@@ -94,8 +94,9 @@ typedef struct {
 
 typedef struct {
   rawSeq* seq;     /* The start of the sequences */
-  size_t pos;      /* The position where reading stopped. <= size. */
-  size_t posInSequence; /* The position to start at within the sequence when starting a new block */
+  size_t pos;      /* The index in this seqStore where reading stopped. <= size. */
+  size_t posInSequence; /* The position within the rawSeq at index 'pos' where reading
+                           stopped. */
   size_t size;     /* The number of sequences. <= capacity. */
   size_t capacity; /* The capacity starting from `seq` pointer */
 } rawSeqStore_t;