## Context

Using the long range matchfinder (`--long`) with high compression levels (16 and above) currently can cause regressions compared to not using the long mode at all. This is because enabling long mode will cause the matches found by the long range matchfinder to be blindly accepted. This PR changes that behavior, and makes each LDM found a candidate for the optimal parser instead.
 
## --patch-from benchmarks
This PR improves `--long` mode compression the most when `--long` mode is actually useful. `--patch-from` is a great example of `--long` mode being useful, so there are some nice benefits here. With this PR, `--patch-from` becomes even more competitive with `bsdiff` for generating small patches. The following results are fastest of three runs. 

**linux kernels tarballs (5.8 -> 5.9-rc7) : ~1GB dst**
| Name               | Conf    | Patch Size   | % of new file | Peak Mem | CPU% | Total Time |
|--------------------|---------|--------|-------|----------|------|------------|
| `zstd-dev`           | -19 -T0 | 5.90MB | 0.58% | 3.24 GB  | 119% | 3m:36s     |
| `zstd-dev`           | -22 -T0 | 6.55MB | 0.65% | 4.09 GB  | 99%  | 5m:54s     |
| `bsdiff`             | default | 5.18MB | 0.51% | 15.58 GB | 99%  | 12m:08s    |
| `zstd-ldm_optimized` | -19 -T0 | 5.10MB | 0.51% | 3.24 GB  | 162% | 4m:30s     |
| `zstd-ldm_optimized` | -22 -T0 | 4.26MB | 0.42% | 4.09 GB  | 99%  | 9m:20s     |

**distant linux kernels tarballs (4.4.238 -> 5.9-rc7): ~1GB dst**
| Name               | Conf    | Patch Size | % of new file | Peak Mem | CPU% | Total Time |
|--------------------|---------|------------|---------------|----------|------|------------|
| `zstd-dev`           | -19 -T0 | 66.90MB    | 6.62%         | 2.68 GB  | 190% | 3m:38s     |
| `zstd-dev`           | -22 -T0 | 71.85MB    | 7.11%         | 3.53 GB  | 99%  | 9m:05s     |
| `bsdiff`             | default | 111.69MB   | 11.04%        | 10.3 GB  | 99%  | 10m:01s    |
| `zstd-ldm_optimized` | -19 -T0 | 58.84MB    | 5.82%         | 2.68 GB  | 220% | 3m:49s     |
| `zstd-ldm_optimized` | -22 -T0 | 56.54MB    | 5.59%         | 3.53 GB  | 99%  | 12m:34s    |

**wordpress.tar (5.5 -> 5.5.1): ~80MB dst**
| Name               | Conf    | Patch Size | % of new file | Peak Mem | CPU% | Total Time |
|--------------------|---------|------------|---------------|----------|------|------------|
| `zstd-dev`           | -19 -T0 | 21.31KB    | 0.04%         | 174 MB   | 99%  | 9.8s       |
| `bsdiff`             | default | 13.67KB    | 0.03%         | 800 MB   | 99%  | 23.8s      |
| `zstd-ldm_optimized` | -19 -T0 | 18.94KB    | 0.04%         | 174 MB   | 99%  | 9.8s       |

**zstd.tar (1.4.4 -> 1.4.5): ~7MB dst**
| Name               | Conf    | Patch Size | % of new file | Peak Mem | CPU% | Total Time |
|--------------------|---------|------------|---------------|----------|------|------------|
| `zstd-dev`           | -19 -T0 | 91.13KB    | 1.12%         | 95 MB    | 99%  | 1.6s       |
| `bsdiff`             | default | 124.23KB   | 1.60%         | 83 MB    | 99%  | 2.3s       |
| `zstd-ldm_optimized` | -19 -T0 | 86.74KB    | 1.12%         | 95 MB    | 99%  | 1.6s       |

Of note is that using `-22 --ultra -T0` with `--patch-from` on actually can cause compressed size regressions on `dev`, compared to `-19 -T0`. This issue is no longer present this PR's branch, however.

## General compression benchmarks

The original goal of this PR was to show that there are no compressed size regression with using `--long` with the opt parser. We demonstrate that fact with the following benchmarks, and show that we can actually get some nice improvements in compression ratio too, particularly in scenarios where `--long` mode is meant to work well. The following results are the fastest of 5 runs.

**five_linux_kernels.tar (4.19.148, 5.4.68, 5.8.12, 5.9-rc7, 5.8.12) ~5GB = 5259889920 bytes**
| Name         | Conf                   | Size      | % of original | CPU% | Time    |
|--------------|------------------------|-----------|-----------------|------|---------|
| `zstd-dev` | -22 --zstd=wlog=30 -T0 | 232896847 | 4.43% | 471% | 14m:30s |
| `zstd-ldm_opt` | -22 --zstd=wlog=30 -T0 | 232896847 | 4.43% | 470% | 14m:31s |
| `zstd-dev` | -22 --long=30 -T0   | 185109024 | 3.52% | 340% | 9m:50s |
| `zstd-ldm_opt` |-22 --long=30 -T0 | 154324523 | 2.93% | 411%  | 12m:30s |

**two_linux_kernels.tar (5.8, 5.9-rc7): ~2GB**
| Name         | Conf                   | Size      | % of original | CPU% | Time    |
|--------------|------------------------|-----------|-----------------|------|---------|
| `zstd-dev`     | -22 --zstd=wlog=30 -T0 | 131944438 | 6.61%           | 195% | 10m:46s |
| `zstd-ldm_opt` | -22 --zstd=wlog=30 -T0 | 131944438 | 6.61%           | 195% | 10m:40s |
| `zstd-dev`     | -22 --long=30 -T0      | 123668734 | 6.20%           | 124% | 9m:32s  |
| `zstd-ldm_opt` | -22 --long=30 -T0      | 120585499 | 6.06%           | 163% | 10m:13s |
| `xz`           | -9 --extreme -T0       | 113720040 | 11.24%          | 951% | 3m:08s  |

`xz` at highest compression settings included here just for fun - it also uses ~8 GB memory compared to 4 GB of `zstd -22 --long=30` too. But it parallelizes the workload incredibly well.

**silesia.tar: ~200MB**

| Name         | Conf                   | Size     | Pct of original | CPU% | Time   |
|--------------|------------------------|----------|-----------------|------|--------|
| `zstd-dev`     | -22 --zstd=wlog=27 -T0 | 52701591 | 24.87%         | 100% | 98.23s |
| `zstd-ldm_opt` | -22 --zstd=wlog=27 -T0 | 52701591 |24.87%          | 100% | 98.87s |
| `zstd-dev`     | -22 --long=27 -T0      | 52777588 | 24.90%          | 100% | 102.23s |
| `zstd-ldm_opt` | -22 --long=27 -T0      | 52695282 | 24.86%          | 100% | 97.98s |

| Name         | Conf                   | Size     | Pct of original | CPU% | Time   |
|--------------|------------------------|----------|-----------------|------|--------|
| `zstd-dev`     | -19 --zstd=wlog=27 -T0 | 53058292 | 25.03%          | 100% | 74.78s |
| `zstd-ldm_opt` | -19 --zstd=wlog=27 -T0 | 53058292 | 25.03%          | 100% | 75.44s |
| `zstd-dev`     | -19 --long=27 -T0      | 53307862 | 25.15%          | 100% | 76.87s |
| `zstd-ldm_opt` | -19 --long=27 -T0      | 53032764 | 25.02%          | 100% | 74.42s |

| Name         | Conf                   | Size     | Pct of original | CPU% | Time   |
|--------------|------------------------|----------|-----------------|------|--------|
| `zstd-dev`     | -16 --zstd=wlog=23 -T0 | 55497790 | 26.18%          | 520% | 11.95s |
| `zstd-ldm_opt` | -16 --zstd=wlog=23 -T0 | 55497790 | 26.18%          | 520% | 12.07s |
| `zstd-dev`     | -16 --long=23 -T0      | 55892935 | 26.37%          | 284% | 18.01s |
| `zstd-ldm_opt` | -16 --long=23 -T0      | 55454265 | 26.16%          | 293% | 17.32s |

Particularly at higher levels (18+), there is virtually no difference in speed on `silesia.tar` when using `--long`, and a speed _gain_ on the concatenated linux kernels file. Not to mention the compression ratio gain. As such, it makes a strong argument for enabling `--long` by default on the highest levels.

### Testing

I've added some more unit tests in `playTests.sh` and a unit test in `fuzzer.c` for regressions, although I'm not sure if there might be weird cases where adding a long match candidate to the optimal parser would legitimately cause a compression ratio regression (and as a result, we shouldn't have these regression tests). Either way, it seems like ldm is already tested decently in `fuzzer.c`, so I don't think adding too many additional tests to verify correctness of compression/roundtrip-ability with ldm here would be that helpful. Letting this patch sit in a fuzzer for a bit might be a good idea though.