@senhuang42 2020-10-05
I've made a few follow-up adjustments to improve compression ratio further in these last few commits suggested by @Cyan4973 : inclusion of mid-flight LDMs (`mid_flight`) and removing the `minMatch` adjustment in the ldm generator (`minmatch_rm`). I've included some benchmarks (fastest of two runs) for the additional changes compared to the current branch. On `five_linux_kernels.tar` we notice a _massive_ improvement in compression ratio with the two changes.

I noticed that `silesia.tar` was actually seeing some _tiny_ regressions in compressed size because of these two commits. I suspected that inclusion of LDMs with the same `mlen` but larger/"worse" `offCode` in the candidate `matches` array may have been hurting the ratio, even though they'd be respecting the ordering of the candidate `matches`. Perhaps because the `offCode` of LDMs are almost always huge by a large margin and likely less compressible. As a result, even with a match length equaling the current longest match, it probably doesn't belong in the 2nd or 3rd spot from the end in the `matches` array. 

I removed inclusion of these LDMs in another commit (`bubble_insertion_rm`) and compression ratio improved, particularly on `silesia.tar`, though just on the scale of a few KB.

**five_linux_kernels.tar** on `ldm_optimized`
conf | commit | size | ratio | time | cpu% |
| --- | --- | --- | --- | --- | --- |
| -22 **--long=30** | `HEAD` | 176535234 | 3.36% | 12m:20s | 389% |
| -22 **--long=30** | `+mid_flight` | 174945548 | 3.33% | 11m:39s | 391% | 
| -22 **--long=30** | `+mid_flight+minmatch_rm` | 154797141 | 2.94% | 12m:01s | 417% | 
| -22 **--long=30** | `+mid_flight+minmatch_rm+bubble_insertion_rm `| 154279903 | 2.93% | 11m:57s | 413% | 
|||||||

**silesia.tar** on `ldm_optimized`
conf | commit | size | ratio | time | cpu% |
| --- | --- | --- | --- | --- | --- |
| -22 **--long=27** | `HEAD` | 52699843 | 24.86% | 105.82s | 100% | 
| -22 **--long=27** | `+mid_flight` | 52700110 | 24.86% | 102.66s | 100% | 
| -22 **--long=27** | `+mid_flight+minmatch_rm` | 52714246 | 24.87% | 104.98s | 100% | 
| -22 **--long=27** | `+mid_flight+minmatch_rm+bubble_insertion_rm `| 52695409 | 24.86% | 114.33s | 100% | 
|||||||