@Cyan4973 2020-10-05
> I suspected that inclusion of LDMs with the same mlen but larger/"worse" offCode in the candidate matches array may have been hurting the ratio

The optimal parser presumes that the list of candidates
comes using ___strictly__ increasing_ match lengths.
Hence, adding a candidate with the same length but larger offset
is violating this assumption, and likely results in the parser making less optimal choices.