@Cyan4973 2020-10-05
Note : these improvements may also positively impact the `--patch-from` mode.
Would be interesting to compare with performance tests completed for the `v1.4.5` announcement of `--patch-from`.