@Cyan4973 2020-10-08
This all looks good to me.

I also went ahead, and decided to measure the impact of this PR on `--patch-from` at high compression level (`-19`),
using 2 linux tarball (5.7->5.8) compressed as delta.

And indeed, compression ratio is significantly improved : 
current `dev` branch produces a patch of ~7.33 MB,
while this PR produces a smaller patch of ~6.31 MB.
This is > 10% additional saving, so not bad!

I tried to compare with `bsdiff` compression ratio, but that's one of those examples where `--patch-from` shines for other reasons : it completed the patch in 5mn. On the other hand, `bsdiff` has been running for 1h30mn, and finally failed, primarily due to gigantic memory requirements to produce the diff (> 16 GB). In contrast, `zstd` was fine with 1.3 GB. Still a lot, but nonetheless considerably better.